chore(server): moving things
