+++ packages/twenty-shared/src/constants/QuotedStringRegex.ts
@@ -1,0 +1,1 @@
+export const QUOTED_STRING_REGEX = /^['"](.*)['"]$/;
